Add schedule workflow subcommand for vineyardctl (#1233)
What do these changes do?
-------------------------

As titled.

---------
